做了好多调整

+++ src/repository/image.go
@@ -31,7 +31,7 @@ func (p *Image) TableName() string {
 
 type ImageRepository interface {
 	FindByPostIdLast(postId int64) (res *Image, err error)
-	FindByPostIds(ids []uint) (res []*Image, err error)
+	FindByPostIds(ids []int64) (res []*Image, err error)
 	AddImage(img *Image) error
 	ExistsImageByMd5(val string) bool
 }
@@ -52,7 +52,7 @@ func (c *image) FindByPostIdLast(postId int64) (res *Image, err error) {
 	return
 }
 
-func (c *image) FindByPostIds(ids []uint) (res []*Image, err error) {
+func (c *image) FindByPostIds(ids []int64) (res []*Image, err error) {
 	if err = c.db.Raw("SELECT image_name,image_path,MAX(id) id,post_id,real_path FROM `images`  WHERE `images`.`deleted_at` IS NULL AND ((post_id in (?))) GROUP BY post_id ORDER BY image_time DESC", ids).
 		Scan(&res).Error; err != nil {
 		return
